requirements.txt
--find-links https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu
ultralytics==8.1.37
fastapi==0.111.0
uvicorn==0.29.0
pillow==10.3.0
numpy==1.24.4
pydantic==2.7.1
